==> 05-results.Rmd <==
# Results

```{r read in data_chapter5, include=FALSE}
library(tidyverse)
library(zoo)
library(lubridate)
library(patchwork)

crime17 <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/crime_17.csv", show_col_types = FALSE)
crime18 <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/crime_18.csv", show_col_types = FALSE)
crime19 <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/crime_19.csv", show_col_types = FALSE)
crime20 <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/crime_20.csv", show_col_types = FALSE)
crime21 <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/crime_21.csv", show_col_types = FALSE)
covid <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/cases-by-day.csv", show_col_types = FALSE)
park_crime <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/park_crime.csv", show_col_types = FALSE)
covid_boro <- read_csv("https://raw.githubusercontent.com/KecheZhu/crimedata/main/data/by-boro.csv", show_col_types = FALSE)

crime_all <- rbind(crime17, crime18, crime19, crime20, crime21)

crime17$date <- as.Date(crime17$date, format="%m/%d/%Y")
crime18$date <- as.Date(crime18$date, format="%m/%d/%Y")
crime19$date <- as.Date(crime19$date, format="%m/%d/%Y")
crime20$date <- as.Date(crime20$date, format="%m/%d/%Y")
crime21$date <- as.Date(crime21$date, format="%m/%d/%Y")
covid$date_of_interest <- as.Date(covid$date_of_interest, format="%m/%d/%Y")
```

## Annual Pattern of Major Crime Types

Firstly, we wanted to visualize the patterns of three major types of crimes in NYC: Felony, Misdemeanor, and Violation. We plotted the daily crime counts from 2017 to 2021, and we were curious about: What is the annual pattern of different types of crime? Does the pattern fluctuates over years?

```{r time series, fig.width=15}
df <- rbind(crime17,crime18,crime19,crime20,crime21) %>% 
  group_by(date,level) %>% 
  summarise(count = n()) %>% 
  ungroup()

g <- ggplot() + 
  geom_line(df, mapping = aes(date, count, color = level)) + 
  geom_smooth(df[df$level == "FELONY",], mapping = aes(date, count), se = FALSE) +
  geom_smooth(df[df$level == "MISDEMEANOR",], mapping = aes(date, count), se = FALSE) +
  geom_smooth(df[df$level == "VIOLATION",], mapping = aes(date, count), se = FALSE) +
  ggtitle("NYC Crimes") +
  labs (x = "", y = "Count") +
  theme_grey(16) +
  theme(legend.title = element_blank())
g
```

**Observations & Analysis:**

1. There is a clear and stable annual pattern from 2017 to 2019: crime counts first increase and reach the highest level in the middle of a year. Then they gradually decrease to their original level at the beginning of each year. The average level of crime counts is stable over years. 
2. All three types of crime follow the pattern. However, the level of Misdemeanor at the very beginning of each year is dramatically high, and it drops to its normal level within a few days.
3. There is a sharp decline of crime counts around March 2020. Although the annual pattern of "first increase then decrease" still holds after that, the overall level of crime counts has been increasing since then. We think that the pattern shown in 2017 - 2019 is the normal pattern while the fluctuations in 2020 and 2021 might be caused by COVID-19. 
4. The crime counts of Felony reach the highest level in 5 years soon after the steep decrease in March 2020.

## Crime Reports in Relation to Boroughs

We plotted the following bar charts to find clues about the relationship between crime data and different areas.

```{r}
bar_df <- crime_all %>%
  group_by(borough, level, year) %>%
  summarise(count = n()) %>% 
  ungroup()

crime_tb <- bar_df[bar_df$borough == "BRONX", ] %>%
  mutate(area = 109.3) %>%
  mutate(rate = count/area)
crime_b <- bar_df[bar_df$borough == "BROOKLYN", ] %>%
  mutate(area = 179.7) %>%
  mutate(rate = count/area)
crime_m <- bar_df[bar_df$borough == "MANHATTAN", ] %>%
  mutate(area = 58.8) %>%
  mutate(rate = count/area)
crime_q <- bar_df[bar_df$borough == "QUEENS", ] %>%
  mutate(area = 281.5) %>%
  mutate(rate = count/area)
crime_s <- bar_df[bar_df$borough == "STATEN ISLAND", ] %>%
  mutate(area = 148.9) %>%
  mutate(rate = count/area)
bar_df2 <- rbind(crime_tb, crime_b, crime_m, crime_q, crime_s)
```
```{r}
g1 <- ggplot(bar_df, aes(x = reorder(borough, -count), y = count, fill = level)) +
  geom_col(position = "dodge") +
  facet_wrap(~year) +
  ggtitle("NYC Crime Count by Borough, Level & Year") +
  xlab("Borough") +
  ylab("Crime Case Count") +
  guides(fill=guide_legend("Crime Level")) +
  theme(axis.text.x = element_text(size = 4, angle = 50, hjust = 1))

g1
```

**Observations & Analysis:**

1. Over the past 5 years, reported crime cases of NYC's 5 boroughs have the following order: Brooklyn > Manhattan > Bronx > Queens > Staten Island.
2. Crime count shows a decreasing trend in general.
3. The ratio of felony cases with respect to total cases has been increasing since 2017.

```{r}
g2 <- ggplot(bar_df2, aes(x = reorder(borough, -rate), y = rate, fill = level)) +
  geom_col(position = "dodge") +
  facet_wrap(~year) +
  ggtitle("NYC Crime Rates by Borough, Level & Year, Adjusted by Area (sq. km)") +
  xlab("Borough") +
  ylab("Crime Rate") +
  guides(fill=guide_legend("Crime Level")) +
  theme(axis.text.x = element_text(size = 4, angle = 50, hjust = 1))

g2
```

**Observations & Analysis:**

1. The crime rate displays a slightly different order: Manhattan > Bronx > Brooklyn > Queens > Staten Island. Manhattan becomes the most crime-dense area, with crime rate largely surpassing every other borough since 2017.
2. We can still observe a decreasing trend, which is consistent with the previous graph.

An interesting observation is that the number of felony reports remained relatively stable over the past five years (compared to the other two major levels). Thus, the ratio of felony cases with respect to total counts has shown an uprising pattern, in contrast to the trend of total crime counts.

## Crime Counts vs COVID-19 Cases

Here we plotted the daily crime counts in 2020 and 2021 together with COVID daily new cases in NYC to explore the impacts of COVID-19 on crime counts.

```{r covid impacts, fig.width=15}
df <- rbind(crime20, crime21) %>% 
  group_by(date,level) %>% 
  summarise(count = n()) %>% 
  ungroup()

ggplot() + 
  geom_line(df, mapping = aes(date, count, color = level)) +
  geom_line(covid, mapping = aes(date_of_interest, CASE_COUNT/5, linetype = "COVID"), color = "darkorchid1") +
  scale_y_continuous("Crime Count", sec.axis = sec_axis(~ (.)*5, name = "Covid Count")) +
  ggtitle("NYC Crimes vs Covid") +
  labs (x = "") +
  theme_grey(16) +
  theme(legend.title = element_blank())
```

We also plotted reductions in crime counts by borough, and compared it with total COVID cases by borough.

```{r}
covid_boro <- covid_boro[-6,]

gc <- ggplot(covid_boro, aes(x = BOROUGH_GROUP, y = CONFIRMED_CASE_COUNT)) +
  geom_col(position = "dodge", fill = "#F8766D") +
  xlab("Borough") +
  ylab("Confirmed Case Count") +
  ggtitle("NYC Total Covid Cases by Borough, 2020-2021")

crime_c20 <- crime_all[crime_all$year >= 2020, ]
crime_f20 <- crime_c20[crime_c20$level == "FELONY", ] %>%
  group_by(borough) %>%
  summarize(count = n()) %>%
  mutate(ann = count/2) %>%
  ungroup()
crime_s20 <- crime_c20 %>%
  group_by(borough) %>%
  summarize(count = n()) %>%
  mutate(ann = count/2) %>%
  ungroup()

crime_c17 <- crime_all[crime_all$year < 2020, ]
crime_f17 <- crime_c17[crime_c17$level == "FELONY", ] %>%
  group_by(borough) %>%
  summarize(count = n()) %>%
  mutate(ann = count/3) %>%
  ungroup()
crime_s17 <- crime_c17 %>%
  group_by(borough) %>%
  summarize(count = n()) %>%
  mutate(ann = count/3) %>%
  ungroup()

crime_mf <- merge(crime_f20, crime_f17, "borough") %>%
  mutate(cate = "Felony")
crime_mt <- merge(crime_s20, crime_s17, "borough") %>%
  mutate(cate = "Total")
crime_red <- rbind(crime_mt, crime_mf) %>%
  mutate(dif = ann.y - ann.x)

gc <- ggplot(covid_boro, aes(x = BOROUGH_GROUP, y = CONFIRMED_CASE_COUNT)) +
  geom_col(position = "dodge", fill = "#F8766D") +
  xlab("Borough") +
  ylab("Confirmed Case Count") +
  ggtitle("NYC Total Covid Cases by Borough") +
  theme(axis.text.x = element_text(size = 4))

gbar <- ggplot(crime_red, aes(x = borough, y = dif, fill = cate)) +
  geom_col(position = "identity") +
  ggtitle("Decrease in NYC Average Annual Crime Counts (Before/After 2020)") +
  xlab("Borough") +
  ylab("Crime Counts") +
  guides(fill=guide_legend("Count")) +
  theme(axis.text.x = element_text(size = 4))
```

```{r}
gc + gbar
```

**Observations & Analysis:**

The steep decrease in crime counts happens soon after COVID-19 breaks out in NYC. As COVID daily new cases reach the first peak and start to decrease, crime counts raise up. However, as COVID-19 breaks out again in 2021, there is little fluctuation in crime counts. 
Still, COVID has contributed to lower crime counts in the long run. The bar charts show that annual crime counts has dropped by large numbers since the beginning of 2020. Further, excluding Queens, reductions in felony & total crime counts are approximately proportionate to total COVID cases. 
Our analysis is that: human activities were severely constrained due to the breakout of COVID-19, which led to the sharp decrease in crime counts. As the situation got better and people became active again, crime counts raised back to their original levels as a result. Since vaccines became available at the beginning of 2021 and people had already adapted to the existence of COVID-19, the second breakout of COVID had little impact on human activities including criminal behaviors. 

## Felony Counts vs COVID-19 Cases

Felony counts has shown the most interesting pattern in the previous graph. Therefore, we decided to dive into the relationship between Felony Counts and COVID Cases.

The plots above could not explain the sharp increase in Felony counts after the first breakout of COVID-19. We continued to analyze the patterns of Felony Counts in different boroughs.

```{r felony in different boroughs, fig.width=10}
df <- crime20 %>% 
  filter(level == 'FELONY') %>%
  group_by(date,borough) %>% 
  summarise(count = n()) %>% 
  ungroup()

covid20 <- covid %>% filter(year(date_of_interest) == 2020)

ggplot() + 
  geom_line(df, mapping = aes(date, count, color = borough)) +
  geom_line(covid20, mapping = aes(date_of_interest, CASE_COUNT/10, linetype = "COVID"), color = "darkorchid1") +
  scale_y_continuous("Felony Count", sec.axis = sec_axis(~ (.)*10, name = "Covid Count")) +
  ggtitle("NYC Felony vs Covid") +
  labs (x = "") +
  theme_grey(16) +
  theme(legend.title = element_blank())

```

**Observations & Analysis:**

From the plot we can see that the breakout of COVID-19 has little impact on the felony counts in Staten island. Although there is an increase in the other 4 boroughs, only the felony counts in Manhattan boom to a dramatically high level. 

```{r}
crime_t17 <- crime_all[crime_all$year < 2020, ]
crime_t17 <- crime_t17[crime_t17$level == "FELONY", ] %>%
  select(borough, type) %>%
  group_by(borough, type) %>%
  summarize(count = n()) %>%
  ungroup() %>%
  mutate(annual = count/3)

crime_t20 <- crime_all[crime_all$year >= 2020, ]
crime_t20 <- crime_t20[crime_t20$level == "FELONY", ] %>%
  select(borough, type) %>%
  group_by(borough, type) %>%
  summarize(count = n()) %>%
  ungroup() %>%
  mutate(annual = count/2)

g1 <- ggplot(crime_t17, aes(x = borough, y = reorder(type, -annual), fill = annual)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "grey85", high = "red", na.value = "black") +
  xlab("Borough") +
  ylab("Felony Type") +
  ggtitle("2017-2019") +
  guides(fill=guide_legend("Count")) +
  theme_bw() +
  theme(axis.text.y = element_text(size = 5), axis.text.x = element_text(size = 5), legend.position = 'none')

g2 <- ggplot(crime_t20, aes(x = borough, y = reorder(type, -annual), fill = annual)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "grey85", high = "red", na.value = "black", breaks=seq(0,16000,4000), limits=c(0, 16000)) +
  xlab("Borough") +
  ylab("") +
  ggtitle("2020-2021") +
  guides(fill=guide_legend("Count")) +
  theme_bw() +
  theme(axis.text.y = element_text(size = 5), axis.text.x = element_text(size = 5))
```

```{r, fig.width=12}
g3 <- g1 + g2
g3 + plot_annotation(title = 'NYC Average Annual Felony Reports by Borough')
```

**Observations & Analysis:**

It is evident from the heat maps that average annual felony reports decreased after 2019. Nonetheless, there has been a big growth in burglary cases, as it became the third mostly reported felony.

```{r felony of different types, fig.width=10}
felony_type <- function(x) {
  if (x %in% c('GRAND LARCENY OF MOTOR VEHICLE', 'ROBBERY', 'BURGLARY', 'GRAND LARCENY', 'MURDER & NON-NEGL. MANSLAUGHTER', 'RAPE', 'FELONY ASSAULT')) {
    return(x)
  } else {
    return('NON-MAJOR FELONY')
  }
}

df <- crime20 %>% 
  filter(level == "FELONY") 

df$type <- unlist(lapply(df$type, felony_type))

df <- df %>%
  group_by(date,type) %>% 
  summarise(count = n()) %>% 
  ungroup()


ggplot() + 
  geom_line(df, mapping = aes(date, count, color = type)) +
  geom_line(covid20, mapping = aes(date_of_interest, CASE_COUNT/10, linetype = "COVID"), color = "darkorchid1") +
  scale_y_continuous("Felony Count", sec.axis = sec_axis(~ (.)*10, name = "Covid Count")) +
  ggtitle("NYC Felony vs Covid") +
  labs (x = "") +
  theme_grey(16) +
  theme(legend.title = element_blank())

```

**Observations & Analysis:**

From the line chart of different felony types we can see more clearly that the sharp increase is mainly contributed by "Burglary".

Synthesizing the findings we have, we conclude that: the breakout of COVID-19 greatly constrained human activities and therefore caused economic difficulties for many people. As a result, some people would commit crimes once the situation got better. Burglary is one of the most straightforward way to get money and Manhattan is the richest borough in NYC. Therefore, most criminals chose to commit burglaries in Manhattan.  

## Park Crime

Park Crime dataset consists of the counts of 7 major felonies that happened in parks and corresponding time and location information. Due to its outdoor nature, we believe that park crime should be highly correlated with seasons. Here we plotted the park crime counts in each quarter.

```{r park crimes-felony-type, fig.width=10}
func <- function(x, y) {
  return(as.yearqtr(paste(x, y, sep = '-')))
}

park_crime$time <- mapply(func, park_crime$YEAR, park_crime$QUARTER)
features <- c("BOROUGH","MURDER","RAPE","ROBBERY","BURGLARY","GRAND LARCENY","GRAND LARCENY OF MOTOR VEHICLE","time") 

df <- park_crime[,features] %>% gather(key = level, value = count, -BOROUGH, -time)

df <- df %>% 
  group_by(time,level) %>% 
  summarise(count = sum(count)) %>% 
  ungroup()

g <- ggplot() + 
  geom_point(df, mapping = aes(time, count, color = level)) +
  geom_line(df, mapping = aes(time, count, color = level)) + 
  ggtitle("NYC Park Crimes") +
  scale_x_discrete(limits = df$time, labels = unlist(lapply(df$time, function(x) format(as.yearqtr(x), "%Y-Q%q")))) +
  labs (x = "", y = "Count") +
  theme_grey(16) +
  theme(legend.title = element_blank(), axis.text.x = element_text(angle = 90))
g
```

```{r park crimes-felony-area, fig.width=15}
df <- park_crime[,features] %>% gather(key = level, value = count, -BOROUGH, -time)

df <- df %>% 
  group_by(time,BOROUGH) %>% 
  summarise(count = sum(count)) %>% 
  ungroup()

g <- ggplot() + 
  geom_point(df, mapping = aes(time, count, color = BOROUGH)) +
  geom_line(df, mapping = aes(time, count, color = BOROUGH)) + 
  ggtitle("NYC Park Crimes") +
  scale_x_discrete(limits = df$time, labels = unlist(lapply(df$time, function(x) format(as.yearqtr(x), "%Y-Q%q")))) +
  labs (x = "", y = "Count") +
  theme_grey(16) +
  theme(legend.title = element_blank(), axis.text.x = element_text(angle = 90))
g
```

**Observations & Analysis:**

1. From the plots we could see that park crime does have a very strong correlation with seasons. The count in the third quarter (summer) is the highest while the count in winter is the lowest, which is aligned with the pattern of human activities. 
2. Grand Larceny contributes to the largest portion of park crime counts and Robbery is the second.
3. Among the 5 boroughs, Manhattan has the highest level of park crime counts.
4. The breakout of COVID-19 also led to a sharp decrease in park crime counts, but was not followed up with a sharp increase. We think the reason is that human activities in many parks were still quite constrained even after the situation became better.
