==> 07-conclusion.Rmd <==
# Conclusion

The data collected have allowed us to answer most of our questions. We observed a clear seasonal trend of crime cases: peaks and troughs come in summers and winters, respectively. We discovered that the patterns and ratios of crime of each type / in each borough remained stable during the past five years, and identified Manhattan as the most crime-dense area in NYC. From the inclusion of COVID data, we also recognized COVID's inhibitory effect on crimes, in both the short run and the long run.

While we are glad to be able to discover patterns of crimes, we are equally concerned about facts beneath these patterns. Reduced human activities in winter explains the seasonal patterns of crime, but what about patterns among the boroughs? What are the reasons that ratios of crimes in each borough could stay relatively stable? Our data offer little insight for these new questions - it would require stricter selections of data and further analysis to answer them.

Moreover, our results enable us to see two sides of the coin named COVID. This everlasting pandemic has brought enormous damage to the society and to people's lives; it is surely surprising to realize that COVID could have positive effects. Although crimes rebounded to the regular level towards ending of the first outburst, average crime counts has dropped on an annual scale. This fact inevitably leads us to ponder on the question that, have we done better in reducing crimes than a virus?

The patterns and ratios of crimes appear to have strong stability, and thus we expect them to stay the same in the following few years. We also expect COVID's effect on crimes to weaken quickly in the future, as people know better about the virus and treatments become more effective. Ultimately, the most important question for us is that should we get accustomed to these crime patterns and live with it, or should we take actions to reduce crimes.
